prevent long wait times for first result in polyhash rev
if the input word width is 32 bit, the signed and unsigned sums will differ by a multiple of 2^32.
this means that if we have multiple solutions in the signed variant, but we start with the unsigned
variant, we will first find the same partial solutions as in the signed variant
until we reach width 32, meaning we get up to 2^32 partial solutions until we throw away
the unsigned variant and start with the signed variant.

instead, we just specifically handle this "common domain" and split the solutions once we reach the wordsize